+++ generate.sh
@@ -1,3 +1,3 @@
 #! /bin/sh
 dotnet build src/CodeGenerator/CodeGenerator.csproj
-dotnet bin/Debug/net6.0/CodeGenerator.dll
+dotnet bin/Debug/net8.0/CodeGenerator.dll